Split the code into the modules
