+++ index.html
@@ -229,6 +229,6 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-<script src = "js/main.js"></script>
+<script type="module" src="js/main.js"></script>
 </body>
 </html>
